Use InputBounds in Openstreetmap
