+++ src/lib.rs
@@ -91,3 +91,51 @@ where
     // data. Please pay attention when using returned data to construct Points
     fn forward(&self, address: &str) -> Result<Vec<Point<T>>, Error>;
 }
+
+/// Used to specify a bounding box to search within when forward-geocoding
+///
+/// - `minimum` refers to the **bottom-left** or **south-west** corner of the bounding box
+/// - `maximum` refers to the **top-right** or **north-east** corner of the bounding box.
+#[derive(Copy, Clone, Debug)]
+pub struct InputBounds<T>
+where
+    T: Float,
+{
+    pub minimum_lonlat: Point<T>,
+    pub maximum_lonlat: Point<T>,
+}
+
+impl<T> InputBounds<T>
+where
+    T: Float,
+{
+    /// Create a new `InputBounds` struct by passing 2 `Point`s defining:
+    /// - minimum (bottom-left) longitude and latitude coordinates
+    /// - maximum (top-right) longitude and latitude coordinates
+    pub fn new<U>(minimum_lonlat: U, maximum_lonlat: U) -> InputBounds<T>
+    where
+        U: Into<Point<T>>,
+    {
+        InputBounds {
+            minimum_lonlat: minimum_lonlat.into(),
+            maximum_lonlat: maximum_lonlat.into(),
+        }
+    }
+}
+
+/// Convert borrowed input bounds into the correct String representation
+impl<T> From<InputBounds<T>> for String
+where
+    T: Float,
+{
+    fn from(ip: InputBounds<T>) -> String {
+        // Return in lon, lat order
+        format!(
+            "{},{},{},{}",
+            ip.minimum_lonlat.x().to_f64().unwrap().to_string(),
+            ip.minimum_lonlat.y().to_f64().unwrap().to_string(),
+            ip.maximum_lonlat.x().to_f64().unwrap().to_string(),
+            ip.maximum_lonlat.y().to_f64().unwrap().to_string()
+        )
+    }
+}
